improves pck settings. fix threads works on msvc. improves js prototypes inherits

+++ source/gui/advisor_chief_window.cpp
@@ -391,7 +391,7 @@ void Chief::Impl::drawMilitary()
           if( minDistance > distance )
           {
             maxThreat = obj;
-            minDistance = distance;
+            minDistance = (int)distance;
           }
         }
       }
